pyproject.toml
[tool.poetry]
name = "cli-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
openai = "^0.28.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^23.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
